+++ src/main/java/googlymod/helpers/GooglyEyeHelpers.java
@@ -5,18 +5,18 @@ import java.util.ArrayList;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 
 public class GooglyEyeHelpers {
-    public static ArrayList<GooglyEye> initEyes(ArrayList<GooglyEyeConfig.CardEye> configs, float drawX, float drawY, float scale) {
+    public static ArrayList<GooglyEye> initEyes(ArrayList<GooglyEyeConfig.CardEye> configs, float originX,float originY, float offsetX, float offsetY, float angle, float scale) {
         ArrayList<GooglyEye> eyes = new ArrayList<>();
         for (GooglyEyeConfig.CardEye config : configs) {
-            eyes.add(new GooglyEye(config, drawX, drawY, scale));
+            eyes.add(new GooglyEye(config, originX, originY, offsetX, offsetY, angle, scale));
         }
         return eyes;
     }
 
-    public static void updateEyes(ArrayList<GooglyEye> eyes, float drawX, float drawY, float scale) {
+    public static void updateEyes(ArrayList<GooglyEye> eyes, float originX,float originY, float offsetX, float offsetY, float angle, float scale) {
         if (eyes == null) return;
         for (GooglyEye eye : eyes) {
-            eye.update(drawX, drawY, scale);
+            eye.update(originX, originY, offsetX, offsetY, angle, scale);
         }
     }