+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.1.0
 * Googly eyes on relics
+* Fix: take card angle into account for eye positions
 
 ### v1.0.0
 * Initial version