+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Googly eyes on relics
 * Googly eyes on character select screen
 * Fix: take card angle into account for eye positions
+* Eye location for mod: Disciple
 
 ### v1.0.0
 * Initial version